wsman-client: fix parameter type in api
Fix warnings like:
openwsman\src\lib\wsman-client.c(653): warning C4244: '=': conversion from 'int' to 'float', possible loss of data

+++ include/wsman-client-api.h
@@ -818,11 +818,11 @@ typedef enum {
 
 
 	void
-	wsmc_set_sub_expiry(int event_subscription_expire, client_opt_t * options);
+	wsmc_set_sub_expiry(float event_subscription_expire, client_opt_t * options);
 
 
 	void
-	wsmc_set_heartbeat_interval(int heartbeat_interval, client_opt_t * options);
+	wsmc_set_heartbeat_interval(float heartbeat_interval, client_opt_t * options);
 
 
 	void
